Move state to parent to prevent more flicker
