+++ src/renderer/components/screens/WelcomeScreen/index.tsx
@@ -1,10 +1,11 @@
-import React from 'react'
+import React, { useLayoutEffect, useState } from 'react'
 
 import Dialog from '../../Dialog'
 import ImageBackdrop from '../../ImageBackdrop'
 import InstantAccountScreen from './InstantAccountScreen'
 import OnboardingScreen from './OnboardingScreen'
 import useInstantOnboarding from '../../../hooks/useInstantOnboarding'
+import { BackendRemote } from '../../../backend-com'
 
 type Props = {
   selectedAccountId: number
@@ -18,18 +19,33 @@ export default function WelcomeScreen({ selectedAccountId, ...props }: Props) {
     showInstantOnboarding,
     switchToInstantOnboarding,
   } = useInstantOnboarding()
+  const [showBackButton, setShowBackButton] = useState(true)
 
   const onClose = () => {
     // Prevent user from closing dialog
   }
+
+  useLayoutEffect(() => {
+    // Show back button when user has already created and configured accounts.
+    // On a fresh DC start we will not have any yet.
+    const checkAccounts = async () => {
+      const accounts = await BackendRemote.listAccounts()
+      if (accounts && accounts.length > 1) {
+        setShowBackButton(true)
+      }
+    }
+
+    checkAccounts()
+  }, [])
 
   return (
     <ImageBackdrop variant='welcome'>
       <Dialog fixed onClose={onClose} width={400}>
         {!showInstantOnboarding ? (
           <OnboardingScreen
-            selectedAccountId={selectedAccountId}
             onNextStep={() => switchToInstantOnboarding(selectedAccountId)}
+            selectedAccountId={selectedAccountId}
+            showBackButton={showBackButton}
             {...props}
           />
         ) : (
